# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Na Pista API",
    "version": "1.0.0",
    "description": "Na Pista — operational infrastructure for businesses (catalogue, stock, orders, services,\nprofessionals, scheduling, appointments), part of the Última Linha ecosystem.\n\n**Authentication:** `Authorization: Bearer <token>` — either a UL Platform user access token (Supabase Auth JWT)\nor a UL Platform service credential (`ulk_…`). See docs/api/authentication.md.\n\n**Tenancy:** every business route is under `/organizations/{organizationId}`; the caller must be an active member\n(human) or hold a credential issued for that organization (service).\n\n**Envelope:** success `{ \"data\": … }` (paginated lists add `pagination`); error\n`{ \"error\": { \"code\", \"message\", \"details\"? } }`. Clients must ignore unknown response fields.\n\n**Authorization extensions:** `x-na-pista-permission` (human role permission), `x-na-pista-scope` (service\ncredential scope; absent = humans only), `x-na-pista-capability` (organization entitlement required)."
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Relative to the Na Pista API host."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/organizations/{organizationId}/services": {
      "post": {
        "operationId": "postOrganizationsByOrganizationIdServices",
        "tags": [
          "Services"
        ],
        "summary": "Create a service",
        "x-na-pista-permission": "services.create",
        "x-na-pista-scope": "catalog.write",
        "x-na-pista-capability": "catalog.enabled",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "durationMinutes": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  },
                  "price": {
                    "type": [
                      "number",
                      "string"
                    ]
                  }
                },
                "required": [
                  "name",
                  "durationMinutes"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Service"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (see `error.details`). Codes: `VALIDATION_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or rejected credentials. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of / not scoped to this organization, missing permission or scope, or capability not enabled. Codes: `FORBIDDEN`, `ENTITLEMENT_REQUIRED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state. Codes: `CONFLICT`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body too large. Codes: `PAYLOAD_TOO_LARGE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error. Report the X-Request-ID. Codes: `INTERNAL_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required dependency (UL Platform or Na Pista's database) is unavailable. Retry later. Codes: `UPSTREAM_UNAVAILABLE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getOrganizationsByOrganizationIdServices",
        "tags": [
          "Services"
        ],
        "summary": "List services",
        "x-na-pista-permission": "services.read",
        "x-na-pista-scope": "catalog.read",
        "x-na-pista-capability": "catalog.enabled",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 50,
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Service"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (see `error.details`). Codes: `VALIDATION_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or rejected credentials. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of / not scoped to this organization, missing permission or scope, or capability not enabled. Codes: `FORBIDDEN`, `ENTITLEMENT_REQUIRED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error. Report the X-Request-ID. Codes: `INTERNAL_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required dependency (UL Platform or Na Pista's database) is unavailable. Retry later. Codes: `UPSTREAM_UNAVAILABLE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organizationId}/services/{serviceId}": {
      "get": {
        "operationId": "getOrganizationsByOrganizationIdServicesByServiceId",
        "tags": [
          "Services"
        ],
        "summary": "Get a service",
        "x-na-pista-permission": "services.read",
        "x-na-pista-scope": "catalog.read",
        "x-na-pista-capability": "catalog.enabled",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "serviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Service"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (see `error.details`). Codes: `VALIDATION_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or rejected credentials. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of / not scoped to this organization, missing permission or scope, or capability not enabled. Codes: `FORBIDDEN`, `ENTITLEMENT_REQUIRED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found in this organization. Codes: `NOT_FOUND`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error. Report the X-Request-ID. Codes: `INTERNAL_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required dependency (UL Platform or Na Pista's database) is unavailable. Retry later. Codes: `UPSTREAM_UNAVAILABLE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patchOrganizationsByOrganizationIdServicesByServiceId",
        "tags": [
          "Services"
        ],
        "summary": "Update, archive or reactivate a service",
        "x-na-pista-permission": "services.update",
        "x-na-pista-scope": "catalog.write",
        "x-na-pista-capability": "catalog.enabled",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "serviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "description": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 2000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "durationMinutes": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  },
                  "price": {
                    "type": [
                      "number",
                      "string",
                      "null"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "ARCHIVED"
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Service"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (see `error.details`). Codes: `VALIDATION_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or rejected credentials. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of / not scoped to this organization, missing permission or scope, or capability not enabled. Codes: `FORBIDDEN`, `ENTITLEMENT_REQUIRED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found in this organization. Codes: `NOT_FOUND`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state. Codes: `CONFLICT`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body too large. Codes: `PAYLOAD_TOO_LARGE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error. Report the X-Request-ID. Codes: `INTERNAL_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required dependency (UL Platform or Na Pista's database) is unavailable. Retry later. Codes: `UPSTREAM_UNAVAILABLE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A UL Platform user access token (JWT) or a UL Platform service credential (`ulk_…`)."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine-readable code. Branch on this, then on the HTTP status."
              },
              "message": {
                "type": "string",
                "description": "Human-readable; not part of the contract (may change)."
              },
              "details": {
                "description": "Present on VALIDATION_ERROR when field-level information exists.",
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                }
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "durationMinutes": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "description": "Exact decimal as a string (money: 2 decimal places; quantities: 6). Never a float."
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "ARCHIVED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 date-time (UTC)."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 date-time (UTC)."
          }
        },
        "required": [
          "id",
          "organizationId",
          "name",
          "description",
          "durationMinutes",
          "price",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ValidationIssue": {
        "type": "object",
        "properties": {
          "location": {
            "type": "string",
            "enum": [
              "body",
              "query",
              "path"
            ]
          },
          "path": {
            "type": "string",
            "description": "Dotted path of the invalid field in the client's input (empty = the whole body)."
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "message"
        ]
      }
    }
  }
}